Move upload data classes to model package
